requirement.txt
aiofiles==0.5.0
aiopg==1.0.0
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
cffi==1.14.2
click==7.1.2
cryptography==2.9.2
dataclasses==0.7
ecdsa==0.14.1
fastapi==0.61.0
h11==0.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.0.0
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
mysqlclient==2.0.1
passlib==1.7.2
peewee==3.13.3
peewee-async==0.7.0
protobuf==3.13.0
psycopg2-binary==2.8.5
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
PyMySQL==0.10.0
python-jose==3.2.0
python-multipart==0.0.5
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
ujson==3.1.0
uvicorn==0.11.8
websockets==8.1
